+++ pkgj-0.28/ci/conan-fmt/test_package/test_package.cpp
@@ -1,8 +1,0 @@
-#include <string>
-#include "fmt/format.h"
-
-int main() {
-    std::string thing("World");
-    fmt::print("Hello {}!\n", thing);
-    return 0;
-}